Modify CleanupDefaultAge (in ConfigurationStruct for notifications and elsewhere) from type int64 to int, plus cascading changes thereby necessitated.

+++ internal/support/notifications/const.go
@@ -44,7 +44,7 @@ type ConfigurationStruct struct {
 	MongoDBKeepAlive              bool
 	ReadMaxLimit                  int
 	ResendLimit                   int
-	CleanupDefaultAge             int64
+	CleanupDefaultAge             int
 	SchedulerNormalDuration       string
 	SchedulerNormalResendDuration string
 	SchedulerCriticalResendDelay  int
